requirements.txt
streamlit>=1.38.0
python-dotenv>=1.0.1
langchain-core>=0.2.25
langchain-openai>=0.1.17
langchain-text-splitters>=0.2.1
PyPDF2>=3.0.1
